Create ants project
This project had an elaborate GUI that let the user play an ants-based Plants vs Zombies knock-off. These are just a few of the files, again it takes a few more and all working in concert to work. Please do not steal this code for homework/projects! While this license does not mention educational use, you'd have to ask a lawyer to be sure such behavior counts as "personal use," and it violates the code of conduct of any university.